rbd: squash ExportDiff.cc into Export.cc
